% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entities.R
\name{average_entity_factors}
\alias{average_entity_factors}
\title{Average entity factors}
\usage{
average_entity_factors(
  scored_entity_data,
  group_col,
  coverages,
  output_prefix = "avg_entity_factor_"
)
}
\description{
Average entity factors
}
