% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entities.R
\name{join_entity_factors}
\alias{join_entity_factors}
\title{Backward-compatible join wrapper}
\usage{
join_entity_factors(rating_data, entity_factor_data, by)
}
\description{
Backward-compatible join wrapper
}
