% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup_factor_value}
\alias{lookup_factor_value}
\title{Lookup a factor value by source}
\usage{
lookup_factor_value(
  row,
  coverage,
  plan,
  term_name,
  value_source = "factor_lookup",
  lookup_var = NULL,
  bounds = "error"
)
}
\description{
Lookup a factor value by source
}
