% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup_interpolated_value}
\alias{lookup_interpolated_value}
\title{Interpolated factor lookup}
\usage{
lookup_interpolated_value(
  row,
  coverage,
  plan,
  term_name,
  lookup_var,
  bounds = "error"
)
}
\description{
Interpolated factor lookup
}
