% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup_term_value}
\alias{lookup_term_value}
\title{Compatibility wrapper for exact term lookup}
\usage{
lookup_term_value(row, coverage, plan, term_name, return_match = FALSE, ...)
}
\description{
Compatibility wrapper for exact term lookup
}
