% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.rating_plan.R
\name{print.rating_plan}
\alias{print.rating_plan}
\title{Print a Rating Plan}
\usage{
\method{print}{rating_plan}(x, ...)
}
\arguments{
\item{x}{A \code{rating_plan} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Prints a compact summary of a \code{rating_plan} object.
}
\examples{
ex <- example_rating_plan()
print(ex$plan)
}
