% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_policies_with_trace.R
\name{rate_policies_with_trace}
\alias{rate_policies_with_trace}
\title{Rate Policies With Trace Output}
\usage{
rate_policies_with_trace(
  policies,
  plan,
  trace_detail = c("terms", "matches", "all"),
  include_inputs = FALSE
)
}
\arguments{
\item{policies}{Policy-level input data frame.}

\item{plan}{A \code{rating_plan} object.}

\item{trace_detail}{One of \code{"terms"}, \code{"matches"}, or \code{"all"}.}

\item{include_inputs}{If \code{TRUE}, include an \code{input_snapshot} table in the
returned object.}
}
\value{
A list of class \code{rating_result} with at least \code{rated_data} and
\code{term_trace}. With \code{trace_detail = "matches"} or \code{"all"}, also includes
\code{factor_match_trace}. With \code{include_inputs = TRUE} or \code{trace_detail = "all"},
also includes \code{input_snapshot}.
}
\description{
Rates policies and returns the rated data plus normalized trace tables.
}
\examples{
ex <- example_rating_plan()
result <- rate_policies_with_trace(ex$policies, ex$plan, trace_detail = "matches")
result$rated_data
result$term_trace
}
