% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helpers.R
\name{rate_set_to_tables}
\alias{rate_set_to_tables}
\title{Split a rate set into review-friendly tables}
\usage{
rate_set_to_tables(factor_table)
}
\description{
Split a rate set into review-friendly tables
}
