% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratingtables-package.R
\docType{package}
\name{ratingtables-package}
\alias{ratingtables}
\alias{ratingtables-package}
\title{ratingtables: Table-driven insurance rating}
\description{
Provides a lightweight, table-driven engine for executing insurance rating plans, including coverage-specific rating specifications, entity aggregation, and trace output for auditing. Given policy data, an ordered rating specification, and rating factor tables, it returns rated policies and, optionally, a step-by-step trace of the calculation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gs-actuary/ratingtables}
  \item Report bugs at \url{https://github.com/gs-actuary/ratingtables/issues}
}

}
\author{
\strong{Maintainer}: Greg Sollenberger \email{yagreg7@yahoo.com} [copyright holder]

Authors:
\itemize{
  \item Greg Sollenberger \email{yagreg7@yahoo.com} [copyright holder]
}

}
\keyword{internal}
