% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{required_policy_fields}
\alias{required_policy_fields}
\title{Required policy fields for a plan}
\usage{
required_policy_fields(plan)
}
\arguments{
\item{plan}{A rating plan.}
}
\value{
A character vector containing the unique policy-data field names
required to execute the rating plan.
}
\description{
Required policy fields for a plan
}
\examples{
ex <- example_rating_plan()

required_policy_fields(ex$plan)
}
