% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_rate_sets}
\alias{validate_rate_sets}
\title{Validate rate set fields}
\usage{
validate_rate_sets(factor_table)
}
\arguments{
\item{factor_table}{A factor table.}
}
\value{
Invisibly returns \code{TRUE} if validation succeeds. Otherwise, the
function stops with an error.
}
\description{
Validate rate set fields
}
\examples{
ex <- example_rating_plan()

isTRUE(
  validate_rate_sets(
    ex$plan$factor_table
  )
)
}
