% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entities.R
\name{aggregate_entity_values}
\alias{aggregate_entity_values}
\title{Aggregate rated entity values to parent/group rows}
\usage{
aggregate_entity_values(
  rated_entity_data,
  group_col,
  value_cols,
  aggregation = "mean",
  weight_col = NULL,
  output_names = NULL,
  output_prefix = NULL
)
}
\description{
Aggregate rated entity values to parent/group rows
}
