% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_helpers.R
\name{append_rating_factors}
\alias{append_rating_factors}
\title{Append wide rating factors to rated data}
\usage{
append_rating_factors(rated_data, term_trace, by = "row_number")
}
\description{
Append wide rating factors to rated data
}
