% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helpers.R
\name{apply_caps}
\alias{apply_caps}
\title{Apply rate caps to indicated premiums}
\usage{
apply_caps(
  rating_data,
  prior_data,
  by,
  coverages,
  max_increase = NULL,
  max_decrease = NULL
)
}
\description{
Apply rate caps to indicated premiums
}
