% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounding.R
\name{apply_rounding}
\alias{apply_rounding}
\title{Apply a rounding rule}
\usage{
apply_rounding(x, rule = NA, digits = NA, increment = NA)
}
\arguments{
\item{x}{Numeric vector.}

\item{rule}{Rounding rule.}

\item{digits}{Digits for base R rounding rules.}

\item{increment}{Increment for nearest_increment.}
}
\description{
Apply a rounding rule
}
\examples{
premiums <- c(101.234, 105.678)

apply_rounding(
  premiums,
  rule = "nearest_cent"
)

apply_rounding(
  premiums,
  rule = "nearest_increment",
  increment = 5
)
}
