% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slot_helpers.R
\name{ensure_slot_columns}
\alias{ensure_slot_columns}
\title{Ensure variable/level slot columns exist}
\usage{
ensure_slot_columns(x, max_vars = 12)
}
\arguments{
\item{x}{Data frame.}

\item{max_vars}{Number of variable/level slots.}
}
\description{
Ensure variable/level slot columns exist
}
\examples{
factors <- data.frame(
  term_name = "territory",
  term_value = 1.10
)

factors <- ensure_slot_columns(
  factors,
  max_vars = 2
)

factors
}
