% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_rating_plan.R
\name{example_rating_plan}
\alias{example_rating_plan}
\title{Build a small example rating plan}
\usage{
example_rating_plan()
}
\description{
Build a small example rating plan
}
\examples{
ex <- example_rating_plan()

ex$policies
ex$plan
}
