% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entities.R
\name{join_entity_values}
\alias{join_entity_values}
\title{Join aggregated entity values back to parent rows}
\usage{
join_entity_values(parent_data, entity_values, by)
}
\description{
Join aggregated entity values back to parent rows
}
