% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slot_helpers.R
\name{make_empty_slots}
\alias{make_empty_slots}
\title{Add empty variable/level slot columns}
\usage{
make_empty_slots(n = 1, max_vars = 12)
}
\arguments{
\item{n}{Number of rows.}

\item{max_vars}{Number of variable/level slots.}
}
\description{
Add empty variable/level slot columns
}
