% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slot_helpers.R
\name{populate_slots}
\alias{populate_slots}
\title{Populate variable/level slots from named values}
\usage{
populate_slots(x, ..., max_vars = 12)
}
\arguments{
\item{x}{Data frame.}

\item{...}{Named vectors/lists of variable names and levels.}

\item{max_vars}{Maximum slot count.}
}
\description{
Populate variable/level slots from named values
}
