% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_core.R
\name{rate_one_row_one_coverage}
\alias{rate_one_row_one_coverage}
\title{Rate one row and one coverage}
\usage{
rate_one_row_one_coverage(row, coverage, plan, row_number = 1)
}
\description{
Rate one row and one coverage
}
