% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_policies.R
\name{rate_policies_with_trace}
\alias{rate_policies_with_trace}
\title{Rate policies and return trace}
\usage{
rate_policies_with_trace(rating_data, plan, validate = TRUE)
}
\description{
Rate policies and return trace
}
