% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helpers.R
\name{rate_set_to_tables}
\alias{rate_set_to_tables}
\title{Split a rate set into review-friendly tables}
\usage{
rate_set_to_tables(factor_table)
}
\arguments{
\item{factor_table}{A normalized long-form factor table.}
}
\value{
A named list of data frames containing review-friendly subsets of
the supplied factor table.
}
\description{
Separate a normalized long-form factor table into a collection of smaller
tables suitable for human review.
}
