% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratingtables-package.R
\docType{package}
\name{ratingtables-package}
\alias{ratingtables}
\alias{ratingtables-package}
\title{ratingtables: Table-driven insurance rating}
\description{
Lightweight table-driven rating kernel for insurance rating tables, coverage-specific rating specifications, interpolation, entity aggregation, custom calculation functions, and trace output.
}
\author{
\strong{Maintainer}: Greg Sollenberger \email{yagreg7@yahoo.com}

Authors:
\itemize{
  \item Greg Sollenberger \email{yagreg7@yahoo.com}
}

}
\keyword{internal}
