% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratingtables-package.R
\docType{package}
\name{ratingtables-package}
\alias{ratingtables}
\alias{ratingtables-package}
\title{ratingtables: Lightweight Table-Driven Rating Engine}
\description{
A dependency-light rating kernel for long-form factor tables, rating specs,
validation, traces, caps, and trace reshaping helpers.
}
\author{
\strong{Maintainer}: Greg Sollenberger \email{yagreg7@yahoo.com}

Authors:
\itemize{
  \item Greg Sollenberger \email{yagreg7@yahoo.com}
}

}
\keyword{internal}
