% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{required_policy_fields}
\alias{required_policy_fields}
\title{Required policy fields for a plan}
\usage{
required_policy_fields(plan)
}
\arguments{
\item{plan}{A rating plan.}
}
\description{
Required policy fields for a plan
}
