% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entities.R
\name{score_entity_rows}
\alias{score_entity_rows}
\title{Backward-compatible entity scoring wrapper}
\usage{
score_entity_rows(entity_data, plan, validate = TRUE)
}
\description{
Backward-compatible entity scoring wrapper
}
