% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_helpers.R
\name{trace_to_excel_style}
\alias{trace_to_excel_style}
\title{Reshape trace to an Excel-style step table}
\usage{
trace_to_excel_style(term_trace)
}
\description{
Reshape trace to an Excel-style step table
}
