% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_helpers.R
\name{trace_to_wide_factors}
\alias{trace_to_wide_factors}
\title{Reshape trace to wide applied values}
\usage{
trace_to_wide_factors(
  term_trace,
  id_cols = c("row_number", "record_id", "coverage")
)
}
\description{
Reshape trace to wide applied values
}
