% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_factor_table}
\alias{validate_factor_table}
\title{Validate a factor table}
\usage{
validate_factor_table(factor_table, max_vars = 12)
}
\arguments{
\item{factor_table}{A factor table.}

\item{max_vars}{Maximum slot count.}
}
\value{
Invisibly returns \code{TRUE} if validation succeeds. Otherwise, the
function stops with an error.
}
\description{
Validate a factor table
}
\examples{
ex <- example_rating_plan()

isTRUE(
  validate_factor_table(
    ex$plan$factor_table,
    max_vars = ex$plan$max_vars
  )
)
}
