% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_policy_data}
\alias{validate_policy_data}
\title{Validate rating input data}
\usage{
validate_policy_data(rating_data, plan)
}
\arguments{
\item{rating_data}{Input data.}

\item{plan}{A rating plan.}
}
\value{
Invisibly returns \code{TRUE} if validation succeeds. Otherwise, the
function stops with an error.
}
\description{
Validate rating input data
}
\examples{
ex <- example_rating_plan()

isTRUE(
  validate_policy_data(
    rating_data = ex$policies,
    plan = ex$plan
  )
)
}
