% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_rating_plan}
\alias{validate_rating_plan}
\title{Validate a rating plan}
\usage{
validate_rating_plan(plan)
}
\arguments{
\item{plan}{A rating plan.}
}
\description{
Validate a rating plan
}
\examples{
ex <- example_rating_plan()

isTRUE(
  validate_rating_plan(ex$plan)
)
}
