% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_rating_spec}
\alias{validate_rating_spec}
\title{Validate a rating specification}
\usage{
validate_rating_spec(rating_spec, ...)
}
\arguments{
\item{rating_spec}{Rating spec.}

\item{...}{Ignored for compatibility.}
}
\description{
Validate a rating specification
}
\examples{
ex <- example_rating_plan()

isTRUE(
  validate_rating_spec(
    ex$plan$rating_spec
  )
)
}
